requirements.txt
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
django-import-export==3.3.5
djangorestframework==3.14.0
Pillow==10.2.0
reportlab==4.0.8
whitenoise==6.6.0
django-admin-interface==0.28.1
django-admin-list-filter-dropdown==1.0.3
mysqlclient==2.2.4
python-dotenv==1.0.1
django-rest-knox==4.2.0
phonenumbers==8.13.37
requests==2.32.3